+++ Model/Document.hpp
@@ -8,20 +8,20 @@ using namespace std;
 class Document
 {
 	public:
-		Document(string _title, string _description = "", float _length = 0);
+		Document(string _title, string _description = "", double _length = 0);
 
 		string GetTitle() const;
 		string GetDescription() const;
-		float GetLength() const;
+		double GetLength() const;
 
 		void SetTitle(string _title);
 		void SetDescription(string _description);
-		void SetLength(float _length);
+		void SetLength(double _length);
 
 	private:
 		string m_title;
 		string m_description;
-		float m_length;
+		double m_length;
 };
 
 #endif
